requirements.txt
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
bert-extractive-summarizer==0.4.2
bleach==3.1.4
blis==0.2.4
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-extensions==2.2.9
django-import-export==2.0.2
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.3
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
plac==0.9.6
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
regex==2020.5.14
requests==2.23.0
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.14.0
spacy==2.1.3
sqlparse==0.3.1
srsly==1.0.2
tablib==1.1.0
terminado==0.8.3
testpath==0.4.4
text-summarizer==0.0.6
thinc==7.0.8
threadpoolctl==2.1.0
tokenizers==0.7.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
transformers==2.10.0
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
